PRODUCTION_AWS+HEROKU/production_simulator/requirements.txt
asgiref==3.2.3
Babel==2.8.0
boto3==1.12.24
botocore==1.15.24
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-bootstrap-form==3.4
django-crispy-forms==1.8.1
django-filter==2.2.0
django-money==1.0
django-phonenumber-field==4.0.0
django-phonenumber-filter==1.0.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
numpy==1.18.1
phonenumbers==8.11.5
Pillow==7.0.0
psycopg2-binary==2.8.4
py-moneyed==0.8.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
